requirements.txt
Flask==2.3.3
pymongo==4.6.1
websocket-client==1.6.4